requirements.txt
appdirs==1.4.3
configparser==3.5.0
enum34==1.1.6
factory_boy==2.8.1
flake8==3.3.0
mccabe==0.6.1
packaging==16.8
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
six==1.10.0
